+++ src/data/DefenseModData.ts
@@ -1,7 +1,9 @@
-import type { ModInterface } from "@/interaces";
-import type OutputModifier from "@/classes/OutputModifier";
+import DamageType from "@/enums/DamageType";
 import ModType from "@/enums/ModType";
 import HasOutputModifier from "@/traits/HasOutputModifier";
+
+import type { ModInterface } from "@/interaces";
+import type OutputModifier from "@/classes/OutputModifier";
 
 export interface DefenseModDataResponse {
     id: string;
@@ -15,6 +17,7 @@ export interface DefenseModDataResponse {
     defenseRangeModifier?: string;
     criticalChanceModifier?: string;
     criticalDamageModifier?: string;
+    elementalAttunement?: string;
     compatibilities?: string;
 }
 
@@ -29,6 +32,7 @@ export default class DefenseModData extends HasOutputModifier implements ModInte
     defenseRange?: OutputModifier;
     criticalChance?: OutputModifier;
     criticalDamage?: OutputModifier;
+    elementalAttunement?: DamageType;
     compatibilities?: string[] = undefined;
     type?: ModType;
 
@@ -52,13 +56,14 @@ export default class DefenseModData extends HasOutputModifier implements ModInte
         this.name = data.name
         this.description = data.description
         this.inTooltip = data.inTooltip
-        this.type = data.type === '' || data.type === undefined ? undefined : ModType[data.type as keyof typeof ModType]
+        this.type = ModType.createEnum(data.type)
         this.defensePower = this.getOutputModifierForValue(data.defensePowerModifier)
         this.defenseHealth = this.getOutputModifierForValue(data.defenseHealthModifier)
         this.defenseRate = this.getOutputModifierForValue(data.defenseRateModifier)
         this.defenseRange = this.getOutputModifierForValue(data.defenseRangeModifier)
         this.criticalChance = this.getOutputModifierForValue(data.criticalChanceModifier)
         this.criticalDamage = this.getOutputModifierForValue(data.criticalDamageModifier)
+        this.elementalAttunement = DamageType.createEnum(data.elementalAttunement)
 
         if (data.compatibilities !== undefined) {
             this.compatibilities = data.compatibilities.split(',')